Add initMap definition to search page for dynamic location marking.
