+++ public/js/navbar-search.js
@@ -1,4 +1,77 @@
-import { handleErrors, api } from './utils.js';
+import { handleErrors, api, addMarkers } from './utils.js';
+window.initMap = (targets) => {
+	const sf = { lat: 37.773, lng: -122.431 };
+	const test2 = { lat: 37.777, lng: -122.418 };
+	const test3 = { lat: 37.779, lng: -122.421 };
+	const test4 = { lat: 37.772, lng: -122.414 };
+
+	//icons
+	const blueFlagIcon = 'http://maps.google.com/mapfiles/ms/icons/flag.png';
+	const purpleIcon = 'http://maps.google.com/mapfiles/ms/icons/purple.png';
+	const purpleIconDot = 'http://maps.google.com/mapfiles/ms/icons/purple-dot.png';
+
+
+	const options = {
+		center: sf,
+		zoom: 13
+	};
+	const map = new google.maps.Map(document.getElementById('map'),
+		options);
+
+
+	const markers = [
+		// { coords: sf, iconImage: purpleIcon, content: '<h3>Resto Info' },
+		// { coords: test2, iconImage: purpleIconDot, content: '<h3>Resto Info' },
+		// { coords: test3, iconImage: blueFlagIcon, content: '<h3>Resto Info' },
+	];
+
+	markers.forEach((props) => {
+		addMarker(props);
+	});
+	// addMarker({
+	// 	coords: sf,
+	// 	iconImage: 'http://maps.google.com/mapfiles/ms/icons/flag.png',
+	// 	content: '<h3>Resto Info'
+	// });
+	// addMarker({
+	// 	coords: test2
+	// });
+	// addMarker({ postion: sf, map: map });
+
+	// const marker = new google.maps.Marker({
+	// 	position: sf,
+	// 	map: map
+	// });
+	// const infoWindow = new google.maps.InfoWindow({
+	// 	content: '<h1>Test Info</h1>'
+	// });
+	// marker.addListener('click', () => {
+	// 	infoWindow.open(map, marker);
+	// })
+
+	function addMarker(props) {
+		const marker = new google.maps.Marker({
+			position: props.coords,
+			map: map,
+		});
+
+		if (props.iconImage) {
+			marker.setIcon(props.iconImage);
+		}
+		//check content
+		if (props.content) {
+			const infoWindow = new google.maps.InfoWindow({
+				content: props.content
+			});
+			marker.addListener('mouseover', () => {
+				infoWindow.open(map, marker);
+			});
+			marker.addListener('mouseout', () => {
+				infoWindow.close(map, marker);
+			})
+		}
+	}
+}
 
 document.addEventListener('DOMContentLoaded', (e) => {
 	// NavBar Selectors
@@ -61,6 +134,11 @@ document.addEventListener('DOMContentLoaded', (e) => {
 		});
 	}
 
+	//add map data for search results
+	function addMapMarkers(businesses) {
+		addMarkers();
+	}
+
 	async function fetchBusinessSearch(body) {
 		try {
 			const res = await fetch(`${api}businesses/search`, {
@@ -76,6 +154,8 @@ document.addEventListener('DOMContentLoaded', (e) => {
 			// declare ratings array to store the ratings
 			generateCards(businesses);
 			findAverageRating(businesses);
+			//addMapMarkers(businesses);
+			window.initMap();
 		} catch (err) {
 			handleErrors(err);
 		}
@@ -114,6 +194,7 @@ document.addEventListener('DOMContentLoaded', (e) => {
 			name: sessionSearchValue
 		};
 		fetchBusinessSearch(body);
+		//window.location.href = '/search';
 	}
 
 	// set up the event listener for the submit button
@@ -128,5 +209,6 @@ document.addEventListener('DOMContentLoaded', (e) => {
 			name: searchValue
 		};
 		fetchBusinessSearch(body);
+		//window.location.href = '/';
 	});
 });
