+++ views/layout.pug
@@ -37,6 +37,7 @@ html
           button.btn.btn-outline-success.my-2.my-sm-0#search-btn(type='submit') Search
           //- TODO: alter search route to be something like /search/?= once we implement /search
     .body-content
+      div#map
       block content
     script(type="module" src="/js/navbar-search.js")
     script(type="module" src="/js/user-auth.js")
